requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
distro==1.9.0
e2b==0.14.0
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonrpcclient==4.0.3
kiwisolver==1.4.5
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langchain-experimental==0.0.50
langsmith==0.0.87
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.11.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
gunicorn